fix: fix: type price

+++ src/shared/infra/typeorm/migrations/1647281677711-ProductTable.ts
@@ -32,7 +32,7 @@ export class ProductTable1647281677711 implements MigrationInterface {
                     },
                     {
                         name: 'price',
-                        type: 'int',
+                        type: 'decimal',
                     },
                     {
                         name: 'creditPoints',
